test(examples): promote tomcat-only assertions into shared bases
The path-variable, request-header, and admin-user tests asserted
starter behavior, not Tomcat behavior, yet only tomcat-mvc ran them.
Move them into the shared bases so jetty-mvc gains them, and add the
request-header assertion to the reactive base for parity. Header names
are compared case-insensitively because servers differ in reported
header-name casing.
