+++ examples/tomcat-mvc/src/test/java/examples/tomcatmvc/BasicAccessLogTest.java
@@ -1,16 +1,11 @@
 package examples.tomcatmvc;
 
 import examples.AbstractBasicAccessLogTest;
-import examples.AccessEventTestUtils;
-import examples.HttpClientTestUtils;
 import io.github.seijikohara.spring.boot.logback.access.LogbackAccessContext;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Basic access log tests for Tomcat + MVC.
@@ -33,31 +28,5 @@ class BasicAccessLogTest extends AbstractBasicAccessLogTest {
     @Override
     protected String getBaseUrl() {
         return "http://localhost:" + port;
-    }
-
-    // Tomcat-specific additional tests
-
-    @Test
-    void getRequestWithPathVariableEmitsAccessEvent() throws Exception {
-        final var response = HttpClientTestUtils.get(getBaseUrl() + "/api/items/42");
-
-        assertThat(response.statusCode()).isEqualTo(200);
-
-        final var events = AccessEventTestUtils.awaitEvents(getListAppender());
-
-        assertThat(events).hasSize(1);
-        final var event = events.get(0);
-        assertThat(event.getRequestURI()).isEqualTo("/api/items/42");
-    }
-
-    @Test
-    void eventCapturesRequestHeaders() throws Exception {
-        HttpClientTestUtils.get(getBaseUrl() + "/api/hello");
-
-        final var events = AccessEventTestUtils.awaitEvents(getListAppender());
-
-        assertThat(events).hasSize(1);
-        final var event = events.get(0);
-        assertThat(event.getRequestHeaderMap()).containsKey("Host");
     }
 }
